--- apps/dockup_ui/test/controllers/page_controller_test.exs
@@ -1,8 +1,0 @@
-defmodule DockupUi.PageControllerTest do
-  use DockupUi.ConnCase
-
-  test "GET /", %{conn: conn} do
-    conn = get conn, "/"
-    assert html_response(conn, 200) =~ "deployments_container"
-  end
-end
