--- apps/dockup_ui/test/views/page_view_test.exs
@@ -1,3 +1,0 @@
-defmodule DockupUi.PageViewTest do
-  use DockupUi.ConnCase, async: true
-end
