+++ apps/dockup_ui/web/services/delete_expired_deployments_service.ex
@@ -13,9 +13,9 @@ defmodule DockupUi.DeleteExpiredDeploymentsService do
     Repo
   }
 
-  def run(service \\ DeleteDeploymentService) do
+  def run(service \\ DeleteDeploymentService, retention_days \\ Dockup.Configs.deployment_retention_days) do
     query = from d in Deployment,
-      where: d.inserted_at < ago(^Dockup.Configs.deployment_retention_days, "day"),
+      where: d.inserted_at < ago(^retention_days, "day"),
       select: d.id
 
     query
